chore: SUI-1848 - refactored E2E tests to make it possible to reuse the Access Token logic
