+++ Apps/Find/tests/SUI.Find.E2ETests/Config.cs
@@ -25,7 +25,17 @@ public record Config
     /// <summary>
     /// Optional ability to override the scope to use when requesting access tokens, to support running the tests through FaUAPI or other gateways.
     /// </summary>
-    public string? AuthScope { get; init; }
+    public string? AuthScope
+    {
+        get;
+        init
+        {
+            field = value;
+            AuthScopes = value != null ? [value] : null;
+        }
+    }
+
+    public string?[]? AuthScopes { get; private init; }
 
     public string? AuthEmulatorHealthCheckEndpoint { get; init; } =
         "https://localhost:7250/api/health";
