+++ Apps/Find/tests/SUI.Find.E2ETests/E2ETestBase.cs
@@ -23,103 +23,4 @@ public class E2ETestBase
             Fixture.StartupDiagnosticMessages == "" ? "None" : Fixture.StartupDiagnosticMessages
         );
     }
-
-    protected async Task<string?> GetAuthTokenAsync(
-        string clientId,
-        string clientSecret,
-        params string?[] scopes
-    )
-    {
-        var originalClientId = clientId;
-
-        clientId =
-            Fixture.Configuration[$"E2E:AuthClientCredentials:{originalClientId}:NewClientId"]
-            ?? clientId;
-
-        clientSecret =
-            Fixture.Configuration[$"E2E:AuthClientCredentials:{originalClientId}:NewClientSecret"]
-            ?? clientSecret;
-
-        var authString = $"{clientId}:{clientSecret}";
-        var base64Auth = Convert.ToBase64String(System.Text.Encoding.UTF8.GetBytes(authString));
-        var clientCredentials = new AuthenticationHeaderValue("Basic", base64Auth);
-
-        return await GetAuthTokenWithRetryAsync(
-            scopes,
-            clientCredentials,
-            clientId,
-            isClientIdSensitive: clientId != originalClientId
-        );
-    }
-
-    private async Task<string> GetAuthTokenWithRetryAsync(
-        string?[] scopes,
-        AuthenticationHeaderValue clientCredentials,
-        string clientId,
-        bool isClientIdSensitive
-    )
-    {
-        const int retryCount = 12;
-        var waitInterval = TimeSpan.FromSeconds(15);
-
-        var retryPolicy = Policy
-            .Handle<Exception>(ex =>
-            {
-                TestOutputHelper.WriteLine(
-                    $"Warning: exception while attempting to get auth token: {ex.Message}"
-                );
-
-                const bool retry = true;
-                return retry;
-            })
-            .WaitAndRetryAsync(
-                retryCount,
-                retryAttempt =>
-                {
-                    TestOutputHelper.WriteLine(
-                        $"Failed to get auth token, waiting for {waitInterval.Seconds} seconds, then retrying, retry {retryAttempt} / {retryCount}..."
-                    );
-                    return waitInterval;
-                }
-            );
-
-        return await retryPolicy.ExecuteAsync(async () =>
-        {
-            var formData = new Dictionary<string, string>
-            {
-                { "grant_type", "client_credentials" },
-                { "scope", string.Join(" ", scopes) },
-            };
-
-            using var content = new FormUrlEncodedContent(formData);
-            using var request = new HttpRequestMessage(
-                HttpMethod.Post,
-                Fixture.Config.AccessTokenUrl
-            );
-
-            request.Content = content;
-            request.Headers.Authorization = clientCredentials;
-
-            TestOutputHelper.WriteLine(
-                $"Requesting access token from: {request.RequestUri} for client ID: {FunctionTestFixture.MaskValue(clientId, isClientIdSensitive)}"
-            );
-
-            var response = await Fixture.Client.SendAsync(request);
-            if (!response.IsSuccessStatusCode)
-            {
-                var error = await response.Content.ReadAsStringAsync();
-                throw new HttpRequestException($"Auth Failed: {response.StatusCode} - {error}");
-            }
-
-            var result = await response.Content.ReadFromJsonAsync<JsonElement>();
-            var accessToken = result.GetProperty("access_token").GetString();
-            if (string.IsNullOrEmpty(accessToken))
-            {
-                throw new InvalidOperationException("No access_token in response");
-            }
-
-            TestOutputHelper.WriteLine("Access token received OK");
-            return accessToken;
-        });
-    }
 }
